Move: common폴더 분리
