+++ src/app/my/_component/ReservationTabContent.tsx
@@ -1,4 +1,4 @@
-import Button from "@/app/_component/common/Button";
+import Button from "@/app/_component/common/atom/Button";
 import React from "react";
 
 const ReservationTabContent = () => {
