+++ src/app/my/_component/UpcomingPackages.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import Button from "@/app/_component/common/Button";
+import Button from "@/app/_component/common/atom/Button";
 
 const UpcomingPackages = () => {
   return (
